examples/user_input/requirements.txt
pygubu-designer
